# pkg/trace/api: improve normalizeTag algorithm