--- a/pkg/trace/api/normalizer_test.go
+++ b/pkg/trace/api/normalizer_test.go
@@ -1,6 +1,7 @@
 package api
 
 import (
+	"bytes"
 	"strings"
 	"testing"
 	"time"
@@ -426,6 +427,20 @@ func TestNormalizeTag(t *testing.T) {
 		{in: "tag:1/2.3", out: "tag:1/2.3"},
 		{in: "---fun:k####y_ta@#g/1_@@#", out: "fun:k_y_ta_g/1"},
 		{in: "AlsO:œ#@ö))œk", out: "also:œ_ö_œk"},
+		{in: strings.Repeat("a", 888), out: strings.Repeat("a", 200)},
+		{
+			in: func() string {
+				b := bytes.NewBufferString("a")
+				for i := 0; i < 799; i++ {
+					_, err := b.WriteRune('🐶')
+					assert.NoError(t, err)
+				}
+				_, err := b.WriteRune('b')
+				assert.NoError(t, err)
+				return b.String()
+			}(),
+			out: "a", // 'b' should have been truncated
+		},
 	} {
 		t.Run("", func(t *testing.T) {
 			assert.Equal(t, tt.out, normalizeTag(tt.in), tt.in)