--- a/releasenotes/notes/apm-improve-normalize-tag-truncation-0ca0061eaf2c9666.yaml
+++ b/releasenotes/notes/apm-improve-normalize-tag-truncation-0ca0061eaf2c9666.yaml
@@ -1,0 +1,12 @@
+# Each section from every releasenote are combined when the
+# CHANGELOG.rst is rendered. So the text needs to be worded so that
+# it does not depend on any information only available in another
+# section. This may mean repeating some details, but each section
+# must be readable independently of the other.
+#
+# Each section note must be formatted as reStructuredText.
+---
+fixes:
+  - |
+    APM: Fixed a bug where normalize tag would not truncate tags correctly
+    in some situations.