This change improves the normalizeTag algorithm to match further
requirements with regards to limiting the maximum tag length allowed.

It also fixes a bug that is encountered when invalid UTF-8 characters are
present.