@gbbr 2019-04-11
Benchmarks compared to original version [pre-refactor](https://github.com/DataDog/datadog-agent/pull/2951) still present significant improvements in speed:
```
benchmark                             old ns/op     new ns/op     delta
BenchmarkNormalizeTag/ok-4            141           68.0          -51.77%
BenchmarkNormalizeTag/trim-4          151           86.6          -42.65%
BenchmarkNormalizeTag/trim-both-4     215           142           -33.95%
BenchmarkNormalizeTag/plenty-4        195           137           -29.74%
BenchmarkNormalizeTag/more-4          316           226           -28.48%

benchmark                             old allocs     new allocs     delta
BenchmarkNormalizeTag/ok-4            2              2              +0.00%
BenchmarkNormalizeTag/trim-4          2              2              +0.00%
BenchmarkNormalizeTag/trim-both-4     3              3              +0.00%
BenchmarkNormalizeTag/plenty-4        3              3              +0.00%
BenchmarkNormalizeTag/more-4          4              4              +0.00%

benchmark                             old bytes     new bytes     delta
BenchmarkNormalizeTag/ok-4            24            24            +0.00%
BenchmarkNormalizeTag/trim-4          32            32            +0.00%
BenchmarkNormalizeTag/trim-both-4     64            64            +0.00%
BenchmarkNormalizeTag/plenty-4        64            64            +0.00%
BenchmarkNormalizeTag/more-4          128           128           +0.00%
```